Perbaiki QuestionController

+++ app/Http/Controllers/API/QuestionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Models\Question;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
@@ -20,9 +21,9 @@ class QuestionController extends Controller
             $questions = Question::search($search)->where('is_status', 1)->get();
 
             // Memuat relasi topics untuk setiap pertanyaan
-            $questions->load('topics');
+            $questions->load('topics', 'user');
         } else {
-            $questions = Question::with('topics')->where('is_status', 1)->get();
+            $questions = Question::with('topics','user')->where('is_status', 1)->get();
         }
 
         // Transformasi hasil untuk mencocokkan format yang Anda inginkan
@@ -63,7 +64,16 @@ class QuestionController extends Controller
 
     public function show(String $slug)
     {
-        $questions = Question::where('slug', $slug)->with('topics')->get();
+        $question = Question::where('slug', $slug)->with('topics', 'user')->first();
+
+        if (is_null($question)) {
+            return response()->json([
+                'status_code' => 404,
+                'status' => 'Error',
+                'message' => 'Data not found',
+                'data' => null,
+            ], 404);
+        }
 
         // Transformasi hasil untuk mencocokkan format yang Anda inginkan
         // $transformedQuestions = $questions->map(function ($question) {
@@ -98,51 +108,51 @@ class QuestionController extends Controller
             'status_code' => 200,
             'status' => 'success',
             'message' => 'Data pertanyaan berhasil diambil',
-            'data' => $questions,
+            'data' => $question,
         ], 200);
     }
 
     public function store(Request $request)
     {
-        $validTopicIds = Rule::exists('topics', 'id');
+        // $validTopicIds = Rule::exists('topics', 'id');
 
         $validation = $request->validate([
-            'question_name' => ['required', 'string'],
-            'question_answer' => ['required'],
-            'topic_id' => ['array', $validTopicIds]
+            'question' => ['required', 'string'],
+            'answer' => ['required'],
+            'topic_id' => ['array', 'exists:topics,id']
         ]);
 
         DB::beginTransaction();
 
-        // try {
-        $question = new Question();
-        $question->user_id = $request->input('user_id');
-        $question->question = $request->input('question_name');
-        $question->slug = $this->generateUniqueSlug($request->input('question_name'));
-        $question->answer = $request->input('question_answer');
-        $question->save();
+        try {
+            $question = new Question();
+            $question->user_id = Auth::user()->id;
+            $question->question = $request->input('question');
+            $question->slug = $this->generateUniqueSlug($request->input('question'));
+            $question->answer = $request->input('answer');
+            $question->save();
 
-        $topicIds = $request->input('topic_id');
-        // Attach each topic to the question
-        $question->topics()->sync($topicIds);
+            $topicIds = $request->input('topic_id');
+            // Attach each topic to the question
+            $question->topics()->sync($topicIds);
 
-        DB::commit();
+            DB::commit();
 
-        return response()->json([
-            'status_code' => 200,
-            'status' => 'success',
-            'message' => 'Data pertanyaan berhasil diambil',
-            'data' => $question,
-        ], 200);
-        // } catch (\Exception $e) {
-        //     DB::rollBack();
+            return response()->json([
+                'status_code' => 200,
+                'status' => 'success',
+                'message' => 'Data pertanyaan berhasil diambil',
+                'data' => $question,
+            ], 200);
+        } catch (\Exception $e) {
+            DB::rollBack();
 
-        //     return response()->json([
-        //         'status_code' => 500,
-        //         'status' => 'Error',
-        //         'message' => 'Data gagal ditambahkan',
-        //     ], 500);
-        // }
+            return response()->json([
+                'status_code' => 500,
+                'status' => 'Error',
+                'message' => 'Data gagal ditambahkan',
+            ], 500);
+        }
     }
 
     private function generateUniqueSlug($question)
@@ -169,12 +179,12 @@ class QuestionController extends Controller
 
     public function update(Request $request, String $slug)
     {
-        $validTopicIds = Rule::exists('topics', 'id');
+        // $validTopicIds = Rule::exists('topics', 'id');
 
         $validation = $request->validate([
-            'question_name' => ['required', 'string'],
-            'question_answer' => ['required'],
-            'topic_id' => ['array', $validTopicIds]
+            'question' => ['required', 'string'],
+            'answer' => ['required'],
+            'topic_id' => ['array', 'exists:topics,id']
         ]);
 
         DB::beginTransaction();
@@ -183,11 +193,20 @@ class QuestionController extends Controller
             // Ambil pertanyaan yang akan diupdate
             $question = Question::where('slug', $slug)->first();
 
+            if (is_null($question)) {
+                return response()->json([
+                    'status_code' => 404,
+                    'status' => 'Error',
+                    'message' => 'Data not found',
+                    'data' => null,
+                ], 404);
+            }
+
             // Update atribut pertanyaan
             $question->update([
-                'question' => $request->input('question_name'),
-                'slug' => $this->generateUniqueSlug($request->input('question_name')),
-                'answer' => $request->input('question_answer'),
+                'question' => $request->input('question'),
+                'slug' => $this->generateUniqueSlug($request->input('question')),
+                'answer' => $request->input('answer'),
             ]);
 
             // Ambil id topik yang baru dari form
@@ -216,9 +235,19 @@ class QuestionController extends Controller
         }
     }
 
-    public function destroy(Question $question)
+    public function destroy(String $slug)
     {
-        $question = Question::where('slug', $question->slug)->first();
+        $question = Question::where('slug', $slug)->first();
+
+        if (is_null($question)) {
+            return response()->json([
+                'status_code' => 404,
+                'status' => 'Error',
+                'message' => 'Data not found',
+                'data' => null,
+            ], 404);
+        }
+
         $question->delete();
 
         return response()->json([
